+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.0.18 - In progress
 ### Added
 - Get favicons on React pages.
+### Changed
+- Upgrade Vitessce to 0.1.3 to fix small bugs and add description for tiff.
 
 ## v0.0.17 - 2020-05-15
 ### Added